Renamed honeypot.ts to cartesi-honeypot.ts Changed Omnichain Layer Zero slug Ref L2B-2951
