+++ packages/config/src/layer2s/cartesi-honeypot.ts
