+++ packages/config/src/layer2s/index.ts
@@ -12,6 +12,7 @@ import { brine } from './brine'
 import { canto } from './canto'
 import { canvasconnect } from './canvasconnect'
 import { capx } from './capx'
+import { honeypot } from './cartesi-honeypot'
 import { HOMEPAGE_MILESTONES } from './common'
 import { degate } from './degate'
 import { degate2 } from './degate2'
@@ -22,7 +23,6 @@ import { fuelv1 } from './fuelv1'
 import { gluon } from './gluon'
 import { grvt } from './grvt'
 import { hermez } from './hermez'
-import { honeypot } from './honeypot'
 import { immutablex } from './immutablex'
 import { immutablezkevm } from './immutablezkevm'
 import { kinto } from './kinto'
